+++ lib/virtualCrypto_web/views/oauth2/balance_view.ex
@@ -1,0 +1,7 @@
+defmodule VirtualCryptoWeb.V1.BalanceView do
+  use VirtualCryptoWeb, :view
+
+  def render("balance.json", %{params: %{data: data}}) do
+    data
+  end
+end
