+++ lib/virtualCrypto_web/views/oauth2/user_view.ex
@@ -9,8 +9,4 @@ defmodule VirtualCryptoWeb.V1.UserView do
       discriminator: params.discriminator
     }
   end
-
-  def render("balance.json", %{params: %{data: data}}) do
-    data
-  end
 end
